Make a lot of adjustments
