+++ src/app/reader/ReaderContainerV2.tsx
@@ -4,6 +4,7 @@ import { useMeasure } from 'react-use'
 import PageContainerV2 from './PageContainerV2'
 import cnBind from 'classnames/bind'
 import style from './reader.css'
+import { Dimensions } from '@/types/dimensions.interface'
 
 const cnJoin = cnBind.bind(style)
 
@@ -12,33 +13,27 @@ export default function ReaderContainerV2(props: {
   current: string
   prev?: string
   next?: string
+  dims: Dimensions
 }) {
-  const [ref, dims] = useMeasure<HTMLDivElement>()
-
   return (
-    <div
-      ref={ref}
-      className={cnJoin(props.className, 'relative overflow-hidden')}
-    >
-      <div className="absolute">
-        {props.prev ? (
-          <PageContainerV2
-            src={props.prev}
-            dimensions={dims}
-            className="absolute z-10 prev"
-          />
-        ) : null}
+    <div className={cnJoin(props.className, 'relative overflow-hidden')}>
+      {props.prev ? (
+        <PageContainerV2
+          src={props.prev}
+          dimensions={props.dims}
+          className="absolute z-10 prev"
+        />
+      ) : null}
 
-        {props.next ? (
-          <PageContainerV2
-            src={props.next}
-            dimensions={dims}
-            className="absolute z-10 next"
-          />
-        ) : null}
+      {props.next ? (
+        <PageContainerV2
+          src={props.next}
+          dimensions={props.dims}
+          className="absolute z-10 next"
+        />
+      ) : null}
 
-        <PageContainerV2 src={props.current} dimensions={dims} />
-      </div>
+      <PageContainerV2 src={props.current} dimensions={props.dims} />
     </div>
   )
 }
