+++ src/app/reader/page.tsx
@@ -1,21 +1,9 @@
-import ReaderContainerV2 from './ReaderContainerV2'
-
-const URLS = [
-  '/placeholder/1.jpg',
-  '/placeholder/2.jpg',
-  '/placeholder/3.jpg',
-  '/placeholder/4.jpg',
-]
+import Wrapper from './Wrapper'
 
 export default function Reader() {
   return (
-    <>
-      <ReaderContainerV2
-        current={URLS[1]}
-        next={URLS[2]}
-        prev={URLS[0]}
-        className="h-screen w-screen"
-      />
-    </>
+    <div>
+      <Wrapper />
+    </div>
   )
 }
